requirements.txt
asgiref==3.3.4
Django==3.2
django-ckeditor==6.1.0
django-debug-toolbar==3.2.1
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
flake8==3.9.1
mccabe==0.6.1
Pillow==8.2.0
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
